requirements.txt
torch
transformers
Pillow
opencv-python
numpy
scikit-learn
imagehash